requirements.txt
# Google Drive API Client & Authentication
pydrive2
google-api-python-client
google-auth-httplib2
google-auth-oauthlib

# Environment variables
python-dotenv

# Retries Logic
tenacity